requirements.txt
selenium
webdriver-manager
beautifulsoup4
lxml
requests
python-dotenv 